Stage 1 of 4 for JetBrains Academy - Kotlin - Number Base Converter project.
Main.kt - requests a decimal number and base from the user and converts it to the requested base, so long as the base is 2, 8, or 16. If the user types in an incorrect base or decimal number the user is informed of their mistake and is requested the number until it is accepted as valid.

README.md - updated to the info for Stage 1