Stage 1 of 4 for JetBrains Academy - Kotlin - Number Base Converter project.
Main.kt - I realized I didn't account for a negative decimal number. I just adjusted it to work like doing toString(radix), which seems to just add a negative to the conversion when the decimal number is negative ( I was thinking toString might do 1's or 2's compliment for base 2, but it doesn't ).